toeic4all-data-generator/pyproject.toml
[tool.poetry]
name = "data4toeic"
version = "0.1.0"
description = ""
authors = ["ruo <comsa333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.22.0"
pyyaml = "^6.0"
sqlalchemy = "^2.0.7"
psycopg2-binary = "^2.9.5"
streamlit-option-menu = "^0.3.2"
stqdm = "^0.0.5"
openai = "^1.16.2"
icecream = "^2.1.3"
langchain = "^0.1.20"
langchain-openai = "^0.1.6"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
